782: Creating an open-banking ingress
Adding an open-banking ingress definition which contains rules that need to be protected by an Open Banking Directory issued OBWac cert.

Moved the /am and /rs rules from the forgerock ingress into the open-banking ingress.

Added a new rule to the forgerock ingress: /am/XUI. When the PSU is accessing the /am/XUI then it is doing consent authorisation via the browser. In which case, AM is simulating an ASPSP (Bank) login portal. Such portals will use the Bank's regular cert, and have nothing to do with open banking.

Currently, this change is just structural, the open-banking ingress is using the same cert as the other ingresses (Let's Encrypt cert) until we have a valid OBWac for our environment.

https://github.com/SecureApiGateway/SecureApiGateway/issues/782