828: Creating test-trusted-directory ingress
This ingress will parse the client certificate if supplied, allowing a combination of calls which use a client cert and those which do not.

The ingress grants access to the /jwkms APIs in IG. It can be removed if a deployment does not want to support the Test Trusted Directory.

https://github.com/SecureApiGateway/SecureApiGateway/issues/828