940 - Add cs_config_manager config for nightly
This config may be used to push v1.0.1 fr-platform-config into our
nightly env. It was create using alpha-1 version of cs_config_manager.

To create this config I did:
- a pull of the nightly config
- ran sort-json on the config
- create an overlay from fr-platform-config commit number 89f88ca
- untarred the overlay
- ran sort-json on the results
- reviewed the changes and dropped any changes that we unnecessary

This config holds a fix for the policy validation errors that were the
cause of issue 940 which resulted in payment submissions being refused
in v1.0.0 of the Secure API Gateway.

In order to push these changes into nightly it will be necessary to:
- Disable the fidc-configurator
- Delete the policy script and the AISP and PISP policies. These were
replaced by the fidc-configurator and have the wrong script ids, so
will need to be deleted or they will prevent the scripts and policies
from being created by the cs_config_manager
- run the cs-config-manager and have it push this config to nightly

Issue: https://github.com/secureapigateway/secureapigateway/issues/940